src/requirements.txt
amqp==2.4.2
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
drf-writable-nested==0.5.1
drf-yasg==1.15.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
kombu==4.5.0
MarkupSafe==1.1.1
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
PyJWT==1.7.1
pylibmc==1.6.0
pyrsistent==0.14.11
pytz==2019.1
requests==2.21.0
ruamel.yaml==0.15.92
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
